+++ lib/src/layout/taskbar.dart
@@ -21,6 +21,7 @@ class ThemedTaskbar extends StatefulWidget {
   final String timeFormat;
   final List<ThemedNavigatorItem> additionalActions;
   final ThemedNavigatorPushFunction? onNavigatorPush;
+  final String? currentPath;
 
   /// [ThemedTaskbar] is the taskbar of the application.
   const ThemedTaskbar({
@@ -91,7 +92,13 @@ class ThemedTaskbar extends StatefulWidget {
     /// [onNavigatorPush] is the callback to be executed when a navigator item is tapped.
     /// By default is `Navigator.of(context).pushNamed`
     this.onNavigatorPush,
+
+    /// [currentPath] is the current path of the navigator. Overrides the default path detection.
+    /// By default, we get the current path from `ModalRoute.of(context)?.settings.name`.
+    this.currentPath,
   });
+
+  static double get height => 55;
 
   @override
   State<ThemedTaskbar> createState() => _ThemedTaskbarState();
@@ -106,12 +113,8 @@ class _ThemedTaskbarState extends State<ThemedTaskbar> with TickerProviderStateM
   List<ThemedNavigatorItem> get items => widget.items;
   List<ThemedNavigatorItem> get persistentItems => widget.persistentItems;
   List<ThemedNotificationItem> get notifications => widget.notifications;
-  double get height => 55;
   ThemedNavigatorPushFunction get onNavigatorPush =>
       widget.onNavigatorPush ?? (path) => Navigator.of(context).pushNamed(path);
-
-  /* 20 = Padding.vertical of main element and 16 = Padding.vertical of the container */
-  double get menuSize => height - (20 + 16);
 
   @override
   void initState() {
@@ -133,7 +136,7 @@ class _ThemedTaskbarState extends State<ThemedTaskbar> with TickerProviderStateM
   @override
   Widget build(BuildContext context) {
     return Container(
-      height: height,
+      height: ThemedTaskbar.height,
       decoration: BoxDecoration(
         color: backgroundColor,
         boxShadow: [
@@ -181,6 +184,7 @@ class _ThemedTaskbarState extends State<ThemedTaskbar> with TickerProviderStateM
                             context: context,
                             backgroundColor: backgroundColor,
                             onNavigatorPush: onNavigatorPush,
+                            currentPath: widget.currentPath,
                           ))
                       .toList(),
                 ),